the u32-ening

+++ idg/rust-viv/src/cli.rs
@@ -10,11 +10,11 @@ use crate::constants;
 pub struct Cli {
     /// Size of the subgrid in pixels
     #[arg(long, default_value = "32")]
-    pub subgrid_size: usize,
+    pub subgrid_size: u32,
 
     /// Size of the grid in pixels
     #[arg(long, default_value = "1024")]
-    pub grid_size: usize,
+    pub grid_size: u32,
 
     /// Length of the observation in hours
     #[arg(long, default_value = "4.0")]
@@ -22,11 +22,11 @@ pub struct Cli {
 
     /// Number of frequency channels
     #[arg(long, default_value = "16")]
-    pub channel_count: usize,
+    pub channel_count: u32,
 
     /// Number of stations
     #[arg(long, default_value = "20")]
-    pub station_count: usize,
+    pub station_count: u32,
 
     /// Starting frequency in hertz
     #[arg(long, default_value = "150e6")]
@@ -57,8 +57,8 @@ pub struct Cli {
 impl Cli {
     // Derived values:
     /// The number of timesteps, as derived from the observation time.
-    pub fn timestep_count(&self) -> usize {
-        (self.observation_hours * 3600.0).floor() as usize
+    pub fn timestep_count(&self) -> u32 {
+        (self.observation_hours * 3600.0).floor() as u32
     }
 
     /// The end frequency, as derived from the start frequency, number of channels, and frequency increment
@@ -72,7 +72,7 @@ impl Cli {
     }
 
     /// The number of baselines, as derived from the number of stations
-    pub fn baseline_count(&self) -> usize {
+    pub fn baseline_count(&self) -> u32 {
         (self.station_count * (self.station_count - 1)) / 2
     }
 }
