+++ idg/rust-viv/src/constants.rs
@@ -1,6 +1,6 @@
 #![allow(unused)]
 
-pub const NR_CORRELATIONS_IN: usize = 2; // XX, YY
-pub const NR_CORRELATIONS_OUT: usize = 1; // I
+pub const NR_CORRELATIONS_IN: u32 = 2; // XX, YY
+pub const NR_CORRELATIONS_OUT: u32 = 1; // I
 pub const W_STEP: f32 = 1.0; // w step in wavelengths
 pub const SPEED_OF_LIGHT: f32 = 299_792_458.0;
